requirements.txt
aws-cdk-lib>=2.109.0
constructs>=10.0.0,<11.0.0
cdk-nag>=2.10.0
bedrock-agent==0.0.7
aws-cdk.aws-lambda-python-alpha